requirements.txt
streamlit==1.19.0
fastapi==0.95.0
uvicorn==0.22.0
requests==2.28.1
pillow==9.4.0
langchain==0.12.0
azure-ai-vision==3.1.0
langgraph==0.1.0
python-dotenv==0.21.0
pytorch==2.0.0
